feat: 8/15-17 现场联调与架构加固（表中心台账/异步命令/日志体系/复盘修复）
- 现场 MCS 联调校准：E37 帧与 W 位、E5 格式码（B=0x21 等）、L 长度=元素数、
  A/U 1 字节长度、SML 显示方言、命名 BUFFER01_02、SCState 固定 2、RESUME→CEID 103
- 表为中心台账：stations 240 行主表 + alarms 清除即删 + device_state 单行；
  第二期异步命令（写表→回 ACK→后台执行→完成/失败刷表，每 PLC 工作线程，重启恢复标记失败）
- 日志体系：五档级别写入日志行、按天滚动保留 7 天、audit 审计通道（SECS 收发摘要/
  命令生命周期/对账）、旁路收编（CommandChannel/Inventory/WebUi/ASP.NET 第三种格式）、
  断线节流、异常堆栈 Debug、帧日志单行折叠、关帧开关省 HEX/SML 转换
- 复盘 A1-A7 修复：T3Ms 配置生效、T3 监控竞态加锁+兜底、事件引擎加锁、
  快照发布深拷贝、cmd_seq 失败回写、9001 下条命令成功自动清（口径已确认）、
  发布 config 自动化（deploy/制作现场linux包.sh，杜绝联调 baseUrl 带到现场）
- 前端：合并台账表/五档日志过滤/日志面板着色/AGVC 与 MCS 联调页/界面 v2 原型
- 测试 106/106 全绿（新增 InventoryTests/LoggingTests/RollingLogTests/
  BypassLoggingTests/FixRegressionTests）；场景脚本 4/4 PASS；linux 包已重发为现场模板
- .gitignore 新增 ~$*（Office 锁文件）；移除误入库的 ~$ 锁文件

+++ tests/BufferC.Harness/McsSim.cs
@@ -12,8 +12,8 @@ public sealed class McsSim : IDisposable
 {
     private TcpClient? _tcp;
     private NetworkStream? _stream;
-    private ushort _sysBytes = 1000;
-    private readonly ConcurrentDictionary<ushort, TaskCompletionSource<HsmsMessage>> _pending = new();
+    private uint _sysBytes = 1000;
+    private readonly ConcurrentDictionary<uint, TaskCompletionSource<HsmsMessage>> _pending = new();
     private readonly ConcurrentQueue<HsmsMessage> _received = new();
     private CancellationTokenSource _cts = new();
 
@@ -54,7 +54,7 @@ public sealed class McsSim : IDisposable
     /// <summary>发送主消息并等待回复</summary>
     public async Task<HsmsMessage?> PrimaryAsync(byte stream, byte func, byte[] body, int timeoutMs = 5000)
     {
-        ushort sys = _sysBytes++;
+        uint sys = _sysBytes++;
         var tcs = new TaskCompletionSource<HsmsMessage>(TaskCreationOptions.RunContinuationsAsynchronously);
         _pending[sys] = tcs;
         var msg = new HsmsMessage { Stream = stream, Function = func, WBit = true, SystemBytes = sys, Body = body };
@@ -179,7 +179,7 @@ public sealed class McsSim : IDisposable
         catch { /* 断连 */ }
     }
 
-    private void SendRaw(byte stream, byte func, bool wbit, byte[] body, ushort sysBytes)
+    private void SendRaw(byte stream, byte func, bool wbit, byte[] body, uint sysBytes)
     {
         try
         {
